requirements.txt
streamlit
torch
torchvision
pillow
scikit-learn
tqdm
matplotlib
pandas
requests
lxml
beautifulsoup4